- Removido Alias pg de pygame para fins de legibilidade - Arquivos refatorados com Black
