+++ main.py
@@ -1,18 +1,18 @@
-import pygame as pg
+import pygame
 from game import Game
 
 
 class Main:
     # Classe basse para o jogo
 
-    pg.init()
+    pygame.init()
 
     def __init__(self):
-        self.window = pg.display.set_mode([600, 750])
-        pg.display.set_caption('Aero')
+        self.window = pygame.display.set_mode([600, 750])
+        pygame.display.set_caption("Aero")
 
         self.loop = True
-        self.fps = pg.time.Clock()
+        self.fps = pygame.time.Clock()
 
         self.game = Game()
         self.lost_gas_length = 0
@@ -22,7 +22,7 @@ class Main:
         self.window.fill([120, 172, 255])
         self.game.draw(self.window)
         self.game.update()
-        pg.draw.rect(self.window, (255, 255, 255), (60, 80, 480, 650), 4)
+        pygame.draw.rect(self.window, (255, 255, 255), (60, 80, 480, 650), 4)
 
     def gas_progress_bar(self):
         # Desenha uma barra de gasolina (ajeitar para atribuir parametros para poder fazer uma barra para o nivel de
@@ -30,17 +30,19 @@ class Main:
         self.game.plane.full_gas += 0.6
         if self.game.plane.full_gas >= 395:
             self.game.plane.full_gas = 395
-        pg.draw.rect(self.window, (255, 0, 0), (10, 80, 25, 400))
-        pg.draw.rect(self.window, (255, 255, 255), (10, 80, 25, 400), 4)
-        pg.draw.rect(self.window, (120, 172, 255), (13, 81, 20, self.game.plane.full_gas))
+        pygame.draw.rect(self.window, (255, 0, 0), (10, 80, 25, 400))
+        pygame.draw.rect(self.window, (255, 255, 255), (10, 80, 25, 400), 4)
+        pygame.draw.rect(
+            self.window, (120, 172, 255), (13, 81, 20, self.game.plane.full_gas)
+        )
 
     def events(self):
         # Identifica todos os eventos na tela
-        for event in pg.event.get():
-            if event.type == pg.QUIT:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
                 self.loop = False
-            if event.type == pg.KEYDOWN:
-                if event.key == pg.K_UP:
+            if event.type == pygame.KEYDOWN:
+                if event.key == pygame.K_UP:
                     self.game.plane.get_gas(50)
             self.game.plane.events(event)
 
@@ -51,7 +53,7 @@ class Main:
             self.events()
             self.fps.tick(60)
             self.gas_progress_bar()
-            pg.display.update()
+            pygame.display.update()
 
 
 Main().update()
